% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lm.R
\name{plot_lm}
\alias{plot_lm}
\title{Plot Residuals from Linear Models}
\usage{
plot_lm(fit, point_size = 1, line_color = "black", line_size = 1,
  binwidth = NULL, hist_fill = "white", hist_color = "black",
  hist_alpha = 1, hist_curve_color = "black", hist_curve_size = 1)
}
\arguments{
\item{fit}{a linear model object produced by lm.}

\item{point_size}{point size for residuals vs fitted plot}

\item{line_color}{line color for residual guides}

\item{line_size}{line size for residual guides}

\item{binwidth}{binwidth for histogram}

\item{hist_fill}{fill color for histogram}

\item{hist_color}{line color for histogram}

\item{hist_alpha}{alpha (transparency) for histogram}

\item{hist_curve_color}{normal curve color for histogram}

\item{hist_curve_size}{normal curve size for histogram}
}
\value{
a gtable of residual plots
}
\description{
produces a grid of three plots: 
 (1) residuals vs predicted values
 (2) normal qq plot
 (3) histogram of residuals with normal curve
}
\examples{
lm_fit <- lm(mpg ~ ., data = mtcars)
plot_lm(lm_fit)

}
