% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_lm.R
\name{print_lm}
\alias{print_lm}
\title{Neat Table of LM Output}
\usage{
print_lm(fit, caption = "LM Results", ci = T, ci_level = 0.95,
  decimals = 4, scipen_val = 999, format = "pandoc",
  full_width = F, align = "r", boot_options = c("striped"))
}
\arguments{
\item{fit}{a linear model object produced by lm.}

\item{caption}{option to add a customized caption to the output table. default caption shows the linear model.}

\item{ci}{option to show confidence intervals for each model parameter.}

\item{ci_level}{option to set the confidence level.}

\item{decimals}{option to set the number of decimals displayed. Must be 1 or more.}

\item{scipen_val}{option to reset scipen.}

\item{format}{A character string: kable format. Possible values are latex, html, markdown, pandoc, and rst. The recommended format is "html" when knitting to .html files, and "latex" when knitting to .pdf files. The pandoc format (default) is recommended when printing to the Console.}

\item{full_width}{Logical: whether html tables should span full width of page.}

\item{align}{Column alignment: a character vector consisting of 'l' (left), 'c' (center) and/or 'r' (right).}

\item{boot_options}{A character vector: arguments to be passed to bootstrap_options for html tables.}
}
\value{
a formatted table of linear model fit results
}
\description{
print_lm() generates neat tables (pandoc, latex, or html formats) of summary output from a linear model object.
}
\examples{
# produce LaTeX table for a .pdf document
lm_fit <- lm(mpg ~ ., data = mtcars)
print_lm(lm_fit)

# produce HTML table for a .html document
lm_fit <- lm(mpg ~ ., data = mtcars)
print_lm(lm_fit, format = "html")

}
