% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_regsub.R
\name{print_regsub}
\alias{print_regsub}
\title{Neat Table of Regsubsets}
\usage{
print_regsub(regfit, format = "data.frame", caption = "Best Subsets",
  bold_crit = T, bold_header = T, generic_header = F, angle = 90,
  full_width = F, font_size = 12, landscape = T, booktabs = T,
  linesep = "", decimals = 4)
}
\arguments{
\item{regfit}{a regsubsets object.}

\item{format}{table format. Must be one of: "data.frame", "latex", "html"}

\item{caption}{option to give the table a caption (html, latex)}

\item{bold_crit}{option to bold the best criterion (html, latex)}

\item{bold_header}{option to bold the headers (html, latex)}

\item{generic_header}{option to use generic headers for the predictors: x1, x2, ...}

\item{angle}{option to align predictor headers at an angle (html, latex)}

\item{full_width}{option to use 100% width (html, latex)}

\item{font_size}{option to set the font size (html, latex)}

\item{landscape}{option to use landscape orientation (html, latex)}

\item{booktabs}{option to use booktabs (html, latex)}

\item{linesep}{option to set line separator (html, latex)}

\item{decimals}{option to set number of decimals}
}
\value{
a data.frame, latex table, or html table
}
\description{
prints a neat table of the a best subsets regression produced by the regsubsets function in the leaps package
}
