% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsq.R
\name{prsq}
\alias{prsq}
\title{Predicted R-Squared from a Linear Regression Model}
\usage{
prsq(fit)
}
\arguments{
\item{fit}{a linear model object produced by lm.}
}
\value{
the predicted r-squared of the linear model
}
\description{
calculates the predicted r-squared from a linear model object produced by lm().
}
\examples{
lm_fit <- lm(mpg ~ ., data = mtcars)
prsq(lm_fit)

}
